Feat: Added new antmatchers and changed Order to house the UUID datatype

+++ src/main/java/com/example/iprwcbackend/controller/ProductController.java
@@ -1,7 +1,12 @@
 package com.example.iprwcbackend.controller;
 
+import com.example.iprwcbackend.dao.AccountDao;
 import com.example.iprwcbackend.dao.ProductDao;
+import com.example.iprwcbackend.model.Account;
+import com.example.iprwcbackend.model.ApiResponse;
 import com.example.iprwcbackend.model.Product;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
@@ -13,9 +18,11 @@ import java.util.List;
 public class ProductController {
 
     private final ProductDao productDao;
+    private final AccountDao accountDao;
 
-    public ProductController(ProductDao productDao) {
+    public ProductController(ProductDao productDao, AccountDao accountDao) {
         this.productDao = productDao;
+        this.accountDao = accountDao;
     }
 
     @GetMapping
@@ -32,19 +39,35 @@ public class ProductController {
 
     @PostMapping
     @ResponseBody
-    public Product addProduct(@RequestBody Product product) {
+    public Object addProduct(@RequestBody Product product) {
+        String email = (String) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        Account account = accountDao.findByEmail(email).get();
+        if(!this.accountDao.isAccountAdmin(account)){
+            return new ApiResponse<>(HttpStatus.UNAUTHORIZED, "You are not authorized to use this.");
+        }
         return productDao.addProduct(product);
     }
 
     @PutMapping("/{id}")
     @ResponseBody
-    public Product updateProduct(@PathVariable Integer id, @RequestBody Product product) {
+    public Object updateProduct(@PathVariable Integer id, @RequestBody Product product) {
+        String email = (String) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        Account account = accountDao.findByEmail(email).get();
+        if(!this.accountDao.isAccountAdmin(account)){
+            return new ApiResponse<>(HttpStatus.UNAUTHORIZED, "You are not authorized to use this.");
+        }
         return productDao.updateProduct(id, product);
     }
 
     @DeleteMapping("/{id}")
     @ResponseBody
-    public void deleteProduct(@PathVariable Integer id) {
+    public Object deleteProduct(@PathVariable Integer id) {
+        String email = (String) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        Account account = accountDao.findByEmail(email).get();
+        if(!this.accountDao.isAccountAdmin(account)){
+            return new ApiResponse<>(HttpStatus.UNAUTHORIZED, "You are not authorized to use this.");
+        }
         productDao.deleteProduct(id);
+        return new ApiResponse<>(HttpStatus.ACCEPTED, "You deleted some data!");
     }
 }
